reast api config ENV plus refactor, wip
