+++ package.json
@@ -8,6 +8,8 @@
     "start": "nf start",
     "start:dev": "webpack-dev-server --config ./config/webpack/app/webpack.config.dev.js",
     "start:prod": "node server",
+    "start:double": "webpack-dev-server --config ./config/webpack/app/webpack.config.dev.js",
+    "start:rest": "cross-env REST_ENV=real webpack-dev-server --config ./config/webpack/app/webpack.config.dev.js",
     "lint": "tslint -c ./tslint.json --force src/**/*.ts src/**/*.tsx",
     "test": "karma start ./config/karma/karma.config.test.js",
     "test:watch": "karma start ./config/karma/karma.config.test.js --no-single-run",
