+++ src/rest-api/login/index.ts
@@ -1,0 +1,8 @@
+export * from './loginAPI.double';
+import { config } from '../config';
+
+if (config.useRealAPI) {
+  console.log('using real api');
+} else {
+  console.log('using double');
+}
